services/manual-input-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-pptx==0.6.21
pika==1.3.2
